Refactor applyRaftCommand by breaking it into two methods.
This eliminates some of the confusion around the nested defer statements and makes
it easier to reason about the response cache and the returned vs. set-in-reply errors.

Removed pre-loaded WriteTooOldError from Range.AddWriteCmd(). It might have been necessary
at one point or another, but there's no remaining justification for it and it might be
responsible for some of the recent issues.

Changed ResponseCache methods to always accept the engine instead of having engine stored
at instantiation. This allows us to set the response cache as part of the same batch as
the executed command.